Fix docs for GET /files/<project_id> endpoint
