+++ docker-app/qfieldcloud/core/serializers.py
@@ -477,3 +477,27 @@ class JobSerializer(serializers.ModelSerializer):
         )
         order_by = "-created_at"
         allow_parallel_jobs = True
+
+
+class FileVersionSerializer(serializers.Serializer):
+    """NOTE not used for actual serialization, but for documentation suing Django Spectacular."""
+
+    size = serializers.IntegerField()
+    md5sum = serializers.CharField()
+    version_id = serializers.CharField()
+    last_modified = serializers.DateTimeField()
+    is_latest = serializers.BooleanField(required=False)
+    display = serializers.CharField()
+    sha256 = serializers.CharField(required=False)
+
+
+class FileSerializer(serializers.Serializer):
+    """NOTE not used for actual serialization, but for documentation suing Django Spectacular."""
+
+    versions = serializers.ListField(child=FileVersionSerializer())
+    sha256 = serializers.CharField(required=False)
+    name = serializers.CharField()
+    size = serializers.IntegerField()
+    md5sum = serializers.CharField()
+    last_modified = serializers.DateTimeField()
+    is_attachment = serializers.BooleanField()
