+++ docker-app/qfieldcloud/core/views/files_views.py
@@ -12,6 +12,7 @@ from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import OpenApiParameter, extend_schema, extend_schema_view
 from qfieldcloud.core import exceptions, permissions_utils, utils
 from qfieldcloud.core.models import Job, ProcessProjectfileJob, Project
+from qfieldcloud.core.serializers import FileSerializer
 from qfieldcloud.core.utils import S3ObjectVersion, get_project_file_with_versions
 from qfieldcloud.core.utils2.audit import LogEntry, audit
 from qfieldcloud.core.utils2.sentry import report_serialization_diff_to_sentry
@@ -39,19 +40,20 @@ class ListFilesViewPermissions(permissions.BasePermission):
         return permissions_utils.can_read_files(request.user, project)
 
 
-class FileVersionSerializer(serializers.Serializer):
-    version_id = serializers.CharField()
-    size = serializers.IntegerField()
-    md5sum = serializers.CharField()
-    last_modified = serializers.DateTimeField()
-    display = serializers.CharField()
-    is_attachment = serializers.BooleanField(required=False)
-
-
 @extend_schema_view(
     get=extend_schema(
         description="Get all the project's file versions",
-        responses={200: serializers.ListSerializer(child=FileVersionSerializer())},
+        responses={200: serializers.ListSerializer(child=FileSerializer())},
+        parameters=[
+            OpenApiParameter(
+                name="skip_metadata",
+                type=OpenApiTypes.INT,
+                required=False,
+                default=0,
+                enum=[1, 0],
+                description="Skip obtaining file metadata (e.g. `sha256`). Makes responses much faster. In the future `skip_metadata=1` might be default behaviour.",
+            ),
+        ],
     ),
 )
 class ListFilesView(views.APIView):
